@backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "LangChain FastAPI backend with Google Gemini"
requires-python = ">=3.9"
dependencies = [
    "fastapi==0.116.1",
    "langchain-google-genai==2.1.3",
    "python-dotenv==1.1.1",
    "uvicorn==0.35.0",
]
